refactor: 组件单独存放一个目录
