--- components/List.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-export default function (props: any) {
-  return (
-    <>
-      <div
-        className={props.className}
-        dangerouslySetInnerHTML={{ __html: props.content }}
-      ></div>
-    </>
-  );
-}
