+++ components/RatingText/index.tsx
