+++ components/SyncProgressModal/index.tsx
